+++ Fire-On-Ice/Assets/Scripts/Magic/FireSword.cs
@@ -37,8 +37,8 @@ public class FireSword : Ability, Cooldown, CastDelay {
 	}
 	
 	void Update (){
-		timeUntilReset = (int)(cooldown - cdTimer + 1f);
-		CooldownUpdate ();
+        timeUntilReset = cooldown - cdTimer + 1f;
+        CooldownUpdate ();
 	}
 	
 	public void CooldownUpdate(){
